Bring back OsuPlayer test
